+++ java-client/src/main/java/io/sphere/client/shop/SignInResult.java
@@ -3,8 +3,11 @@ package io.sphere.client.shop;
 import io.sphere.client.shop.model.Cart;
 import io.sphere.client.shop.model.Customer;
 
-/** Result object returned by {@link io.sphere.client.shop.SphereClient SphereClient's} sign-in and sign-up methods.
- *  Contains a customer and their active cart (if such a cart exists, null otherwise). */
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
+/** Result returned by {@link io.sphere.client.shop.SphereClient SphereClient's} sign-in and sign-up methods.
+ *  Contains a customer and their active cart (if such cart exists, null otherwise). */
 public class SignInResult {
     private Customer customer;
     private Cart cart;
@@ -12,11 +15,8 @@ public class SignInResult {
     // for JSON deserializer
     private SignInResult() { }
 
-    public SignInResult(Customer customer, Cart cart) {
-        this.customer = customer;
-        this.cart = cart;
-    }
-
-    public Customer getCustomer() { return customer; }
-    public Cart getCart() { return cart; }
+    /** The existing customer. */
+    @Nonnull public Customer getCustomer() { return customer; }
+    /** Customer's cart or null if no cart exists. */
+    @Nullable public Cart getCart() { return cart; }
 }